pyproject.toml
[project]
name = "images_api"
authors = [
    {name = "Vlad Naumchik", email = "luckstriker073@gmail.com"},
]
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all]",
    "mangum",
    "pydantic",
    "pymongo",
    "uvicorn",
]


[project.optional-dependencies]
dev = [
    "ipython",

    "pytest-mongodb",
    "pytest-deadfixtures",
    "pytest-env",
    "pytest-freezegun",
    "pytest-mock",
    "pytest-randomly",

    "black",

    "autoflake==1.7.0",

    "freezegun",
    "mixer",

    "flake8-fastapi",
    "flake8-absolute-import",
    "flake8-black",
    "flake8-bugbear",
    "flake8-cognitive-complexity",
    "flake8-eradicate",
    "flake8-isort>=4.0.0",
    "flake8-fixme",
    "flake8-pep3101",
    "flake8-pie",
    "flake8-print",
    "flake8-printf-formatting",
    "flake8-pytest",
    "flake8-pytest-style",
    "flake8-simplify",
    "flake8-todo",
    "flake8-use-fstring",
    "flake8-variables-names",
    "flake8-walrus",
    "flake8-pyproject",

    "mypy",
    "types-freezegun",
    "types-Pillow",
]


[tool.flake8]
max-line-length = 120
inline-quotes = "\""
ignore = [
    "E501",  # Line too long
    "E265",  # Block comments should have one space before the pound sign (#) and the comment itself
    "F811",  # Redefinition of unused name from line n
    "PT001",  # Use @pytest.fixture() over @pytest.fixture
    "SIM102",  # Use a single if-statement instead of nested if-statements
    "SIM113",  # Use enumerate instead of manually incrementing a counter
    "E203",  # whitespace before ':', disabled for black purposes https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices
]
exclude = [
    "frontend",
    ".git",
    "__pycache__",
]


[tool.isort]
profile = "google"
line_length = 120
use_parentheses = true
include_trailing_comma = true
multi_line_output = 3


[tool.black]
line_length = 120
